PMM-7958 Fix PostresSQL db issue
